refactor: extract collections/items parser
